Broke down monolithic component into more bit sized pieces. I might've broken some stuff
